Update README.md
This is an interdisciplinary study investigating to what extent a machine translation (MT) system can reproduce a translator’s style if it is trained on a corpus of previous translations by the same translator. Leech and Short’s (2007) methodology of stylistic analysis was adapted to determine the stylistic features of a prominent Turkish translator, Nihal Yeğinobalı. Qualitative data analysis involved close reading of translations, whereas quantitative analysis was conducted with corpus tools and Python using keyword and morphological data. By comparing the Nihal Yeğinobalı corpus against a reference corpus representative of Turkish literary translations within the same time period, stylistic features were observed on lexical and morphological levels. An authorship attribution analysis showed that the stylistic features can help distinguish between the Nihal Yeğinobalı corpus and the reference corpus. In a following step, the Nihal Yeğinobalı corpus was used to train and test a set of MT models to obtain models fine-tuned to the translator, which were compared with a general-domain pre-trained MT model. The authorship attribution analysis of the MT output resulted in a distinction between the fine-tuned MT model output, attributing this to Nihal Yeğinobalı, and the pre-trained model output, classifying it as reference. In addition, a keyword analysis of the MT output revealed a high degree of similarity in lexical choice between the fine-tuned MT model and Nihal Yeğinobalı. The present findings provide some evidence of style transfer, suggesting that an MT system can reproduce a translator’s style if trained on a corpus of their translations.